requirements.txt
DrissionPage~=4.0.4.4
pillow~=10.2.0
tqdm~=4.66.1
future~=0.18.3
opencv-python~=4.9.0.80